refactor(#54): Payment 구조 변경
